LayoutTests/http/tests/iframe-monitor/dark-mode-expected.txt

CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit.
CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit.
CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit.
Test unloaded HTML supports dark mode correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS setupDone is true
In header, scheme is light and dark.
PASS iframeElem.srcdoc is not ""
PASS window.getComputedStyle(pElem).getPropertyValue('color') is 'rgb(0, 0, 0)'
PASS window.getComputedStyle(pElem).getPropertyValue('color') is 'rgb(255, 255, 255)'
In main, scheme is normal.
PASS iframeElem.srcdoc is not ""
PASS window.getComputedStyle(pElem).getPropertyValue('color') is 'rgb(0, 0, 0)'
PASS window.getComputedStyle(pElem).getPropertyValue('color') is 'rgb(0, 0, 0)'
In footer, scheme is dark only.
PASS iframeElem.srcdoc is not ""
PASS window.getComputedStyle(pElem).getPropertyValue('color') is 'rgb(255, 255, 255)'
PASS window.getComputedStyle(pElem).getPropertyValue('color') is 'rgb(255, 255, 255)'
PASS successfullyParsed is true

TEST COMPLETE
